requirements.txt
python-dotenv>=0.19.0
selenium>=4.0.0
lxml>=4.6.0
requests>=2.25.0